Refactor StreamBIMDownloader to improve FTP handling and path management
- Introduced StreamBimPathHelper for centralized path operations.
- Updated StreamBIMFilesAndFolderAutoFillCollector and StreamBIMProjectRootFolderAutoFillCollector to use new credential provider and FTP client factory.
- Enhanced error handling by consolidating exception types and using a transient failure helper.
- Added StreamBimDownloadService for managing file downloads with retry logic.
- Created StreamBimFileTransferService for handling individual file transfers.
- Implemented StreamBimDownloadOutcomeBuilder to aggregate download results.
- Added StreamBimExceptionHelper for consistent exception message retrieval.
- Introduced StreamBimExtensionResultFactory for creating extension results based on download outcomes.
- Added comprehensive unit tests to ensure reliability of new features and refactorings.
